+++ src/main/java/com/example/final_case_social_web/service/friend_relation/FriendRelationServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.final_case_social_web.repository.FriendRelationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Service
@@ -34,7 +35,7 @@ public class FriendRelationServiceImpl implements IFriendRelationService{
     }
 
     @Override
-    public Iterable<Long> findAllIdUserNotFriend(Long id, Long id1) {
+    public Iterable<BigInteger> findAllIdUserNotFriend(Long id, Long id1) {
        return friendRelationRepository.findIdUserNotFriend(id,id1);
     }
 }
